+++ packages/utils/test/schema/retrieveSchemaTest.ts
@@ -171,6 +171,64 @@ export default function retrieveSchemaTest(testValidator: TestValidatorType) {
       });
     });
 
+    it("should `resolve` and stub out a schema which contains an `additionalProperties` with oneOf", () => {
+      const oneOf: RJSFSchema[] = [
+        {
+          type: "string",
+        },
+        {
+          type: "number",
+        },
+      ];
+      const schema: RJSFSchema = {
+        additionalProperties: {
+          oneOf,
+        },
+        type: "object",
+      };
+
+      const formData = { newKey: {} };
+      expect(retrieveSchema(testValidator, schema, {}, formData)).toEqual({
+        ...schema,
+        properties: {
+          newKey: {
+            type: "object",
+            oneOf,
+            [ADDITIONAL_PROPERTY_FLAG]: true,
+          },
+        },
+      });
+    });
+
+    it("should `resolve` and stub out a schema which contains an `additionalProperties` with anyOf", () => {
+      const anyOf: RJSFSchema[] = [
+        {
+          type: "string",
+        },
+        {
+          type: "number",
+        },
+      ];
+      const schema: RJSFSchema = {
+        additionalProperties: {
+          anyOf,
+        },
+        type: "object",
+      };
+
+      const formData = { newKey: {} };
+      expect(retrieveSchema(testValidator, schema, {}, formData)).toEqual({
+        ...schema,
+        properties: {
+          newKey: {
+            type: "object",
+            anyOf,
+            [ADDITIONAL_PROPERTY_FLAG]: true,
+          },
+        },
+      });
+    });
+
     it("should handle null formData for schema which contains additionalProperties", () => {
       const schema: RJSFSchema = {
         additionalProperties: {